Added a demo of the discrete Fourier transform
This demo is enough to production code, at least in non-GWT code,
except for two things:

First, the URL scheme isn't really ideal.  Currently, the Fourier
transformed tile is accessed exactly the same as the normal tile, except
that device.channel in the URL is replaced with device.channel.DFT.

Second, there is no UI to allow the user to choose whether to view a
spectral view or a normal plot.  The current demo always creates a
spectral view for device PolarStrap, and leaves all other devices
untouched.